+++ hc/api/management/commands/smtpd.py
@@ -7,6 +7,7 @@ from smtpd import SMTPServer
 from django.core.management.base import BaseCommand
 from django.db import connections
 from hc.api.models import Check
+from hc.lib.html import html2text
 
 RE_UUID = re.compile(
     "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-4[a-fA-F0-9]{3}-[8|9|aA|bB][a-fA-F0-9]{3}-[a-fA-F0-9]{12}$"
@@ -22,6 +23,23 @@ def _match(subject, keywords):
     return False
 
 
+def _to_text(message, with_subject, with_body):
+    chunks = []
+    if with_subject:
+        chunks.append(message.get("subject", ""))
+    if with_body:
+        plain_mime_part = message.get_body(("plain",))
+        if plain_mime_part:
+            chunks.append(plain_mime_part.get_content())
+
+        html_mime_part = message.get_body(("html",))
+        if html_mime_part:
+            html = html_mime_part.get_content()
+            chunks.append(html2text(html))
+
+    return "\n".join(chunks)
+
+
 def _process_message(remote_addr, mailfrom, mailto, data):
     to_parts = mailto.split("@")
     code = to_parts[0]
@@ -35,15 +53,16 @@ def _process_message(remote_addr, mailfrom, mailto, data):
         return f"Check not found: {code}"
 
     action = "success"
-    if check.subject or check.subject_fail:
-        action = "ign"
-        # Specify policy, the default policy does not decode encoded headers:
+    if check.filter_subject or check.filter_body:
         data_str = data.decode(errors="replace")
-        parsed = email.message_from_string(data_str, policy=email.policy.SMTP)
-        subject = parsed.get("subject", "")
-        if check.subject_fail and _match(subject, check.subject_fail):
+        # Specify policy, the default policy does not decode encoded headers:
+        message = email.message_from_string(data_str, policy=email.policy.SMTP)
+        text = _to_text(message, check.filter_subject, check.filter_body)
+
+        action = "ign"
+        if check.failure_kw and _match(text, check.failure_kw):
             action = "fail"
-        elif check.subject and _match(subject, check.subject):
+        elif check.success_kw and _match(text, check.success_kw):
             action = "success"
 
     ua = "Email from %s" % mailfrom